requirements.txt
asgiref==3.8.1
certifi==2025.4.26
cffi==1.17.1
charset-normalizer==3.4.2
cryptography==44.0.3
Django==5.2.1
django-allauth==65.8.0
django-tailwind==4.0.1
dotenv==0.9.9
idna==3.10
psycopg2-binary==2.9.10
pycparser==2.22
PyJWT==2.6.0
python-dotenv==1.1.0
requests==2.32.3
sqlparse==0.5.3
tzdata==2025.2
urllib3==2.4.0
langchain==0.3.25
langchain-community==0.3.24
openai==1.78.1
PyMuPDF==1.25.5
sqlalchemy==2.0.41
pandas==2.2.3
pdfplumber==0.11.6
sentence-transformers==4.1.0